requirements.txt
disnake==2.3.0
SQLAlchemy==1.4.29
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mysql-connector-python==8.0.27
python-dotenv==0.19.1
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
yarl==1.7.2
mysqlclient==2.1.0